+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-22T11:07:47.246Z",
+    "2026-04-22T13:05:42.300Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {